+++ crates/factor-key-value-azure/src/lib.rs
@@ -1,5 +1,5 @@
 use serde::Deserialize;
-use spin_factor_key_value::MakeKeyValueStore;
+use spin_factor_key_value::runtime_config::spin::MakeKeyValueStore;
 use spin_key_value_azure::{
     KeyValueAzureCosmos, KeyValueAzureCosmosAuthOptions, KeyValueAzureCosmosRuntimeConfigOptions,
 };
